+++ ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: nappalav <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/09/01 16:51:38 by nappalav          #+#    #+#             */
-/*   Updated: 2023/09/01 17:45:57 by nappalav         ###   ########.fr       */
+/*   Updated: 2023/09/15 22:32:21 by nappalav         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strnstr(const char *haystack, const char *needle, size_t len)
 {
-	int		i;
-	int		j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	if (!needle[i])
